requirements.txt
Flask==1.1.1
matplotlib==3.0.2
numpy==1.20.1
Pillow==7.2.0
imageio==2.9.0
tensorboard==2.2.1
torch==1.4.0+cpu
torchvision==0.5.0+cpu
